requirements.txt
# Delta CFO Agent - Required Dependencies
# Consolidated requirements including invoice processing module

# ============================================
# Core Web Framework
# ============================================
Flask>=2.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0

# ============================================
# Data Processing & Analysis
# ============================================
pandas>=1.3.0
openpyxl>=3.0.0

# ============================================
# AI Integration
# ============================================
anthropic>=0.8.0

# ============================================
# HTTP & Network
# ============================================
requests>=2.25.0

# ============================================
# Date & Time Utilities
# ============================================
python-dateutil>=2.8.0

# ============================================
# Invoice Processing - PDF Processing
# ============================================
PyMuPDF>=1.23.0          # PDF rendering and conversion (fitz)
PyPDF2>=3.0.0            # PDF text extraction
pdfplumber>=0.7.0        # Advanced PDF data extraction
pdf2image>=1.16.0        # PDF to image conversion
reportlab>=3.6.0         # PDF generation for testing

# ============================================
# Invoice Processing - Image Processing
# ============================================
Pillow>=10.0.0           # Image manipulation and format support
opencv-python>=4.8.0     # Advanced image processing

# ============================================
# Invoice Processing - OCR
# ============================================
pytesseract>=0.3.10      # Tesseract OCR wrapper

# ============================================
# Invoice Processing - Email Automation
# ============================================
exchangelib>=4.9.0       # Exchange server integration
imapclient>=2.3.1        # IMAP email client

# ============================================
# File Type Detection & Handling
# ============================================
python-magic>=0.4.27     # File type detection

# ============================================
# Compressed File Support
# ============================================
py7zr>=0.20.0            # 7z archive support

# ============================================
# Database Support - Cloud SQL PostgreSQL
# ============================================
psycopg2-binary>=2.9.9   # PostgreSQL adapter for Cloud SQL
flask-cors>=4.0.0        # CORS support for web APIs

# ============================================
# Google Cloud Dependencies
# ============================================
google-cloud-secret-manager>=2.16.4    # Secret management
cloud-sql-python-connector>=1.4.0      # Cloud SQL connector

# ============================================
# Environment & Configuration
# ============================================
python-dotenv>=1.0.0     # Environment variable management